2.5 / Nodes * fix warnings for msvc

+++ source/blender/editors/space_node/node_header.c
@@ -190,7 +190,7 @@ static void node_make_addmenu(bContext *C, int nodeclass, uiBlock *block)
 				if(ngroup->type==ntree->type) {
 					
 					uiDefIconTextBut(block, BUTM, 1, ICON_BLANK1, (ngroup->id.name+2), 0, 
-						yco-=20, menuwidth, 19, NULL, 0.0, 0.0, 1, NODE_GROUP_MENU+tot, "");
+						yco-=20, menuwidth, 19, NULL, 0.0, 0.0, 1.0f, (float)(NODE_GROUP_MENU+tot), "");
 					a++;
 				}
 			}
@@ -202,11 +202,11 @@ static void node_make_addmenu(bContext *C, int nodeclass, uiBlock *block)
 				if( type->nclass == nodeclass ) {
 					if(type->type == NODE_DYNAMIC) {
 						uiDefIconTextBut(block, BUTM, 1, ICON_BLANK1, type->name, 0, 
-							yco-=20, menuwidth, 19, NULL, 0.0, 0.0, 1, NODE_DYNAMIC_MENU+script, "");
+							yco-=20, menuwidth, 19, NULL, 0.0, 0.0, 1.0f, (float)(NODE_DYNAMIC_MENU+script), "");
 						script++;
 					} else {
 					uiDefIconTextBut(block, BUTM, 1, ICON_BLANK1, type->name, 0, 
-						yco-=20, menuwidth, 19, NULL, 0.0, 0.0, 1, type->type, "");
+						yco-=20, menuwidth, 19, NULL, 0.0, 0.0, 1.0f, (float)(type->type), "");
 					}
 					a++;
 				}
